chore: refactor client config builder code.
