+++ openshift/openshift/src/main/java/org/arquillian/cube/openshift/impl/requirement/OpenshiftRequirement.java
@@ -1,23 +1,21 @@
 package org.arquillian.cube.openshift.impl.requirement;
 
-import io.fabric8.kubernetes.clnt.v3_1.Config;
 import io.fabric8.kubernetes.clnt.v3_1.DefaultKubernetesClient;
 import io.fabric8.kubernetes.clnt.v3_1.KubernetesClient;
 import io.fabric8.kubernetes.clnt.v3_1.utils.URLUtils;
 import io.fabric8.openshift.clnt.v3_1.OpenShiftClient;
 import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.Arrays;
+import java.util.List;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.Response;
-import org.arquillian.cube.openshift.impl.client.CubeOpenShiftConfiguration;
+import org.arquillian.cube.kubernetes.impl.ClientConfigBuilder;
+import org.arquillian.cube.kubernetes.impl.DefaultConfiguration;
+import org.arquillian.cube.kubernetes.impl.ExtensionRegistrar;
 import org.arquillian.cube.spi.requirement.Constraint;
 import org.arquillian.cube.spi.requirement.UnsatisfiedRequirementException;
-import org.jboss.arquillian.config.descriptor.api.ArquillianDescriptor;
-import org.jboss.arquillian.config.impl.extension.ConfigurationRegistrar;
 
-import static org.arquillian.cube.kubernetes.ClientConfigurator.getConfigBuilder;
 import static org.arquillian.cube.kubernetes.impl.DefaultConfigurationFactory.KUBERNETES_EXTENSION_NAME;
 import static org.arquillian.cube.openshift.impl.client.CubeOpenShiftConfigurationFactory.OPENSHIFT_EXTENSION_NAME;
 
@@ -25,7 +23,11 @@ public class OpenshiftRequirement implements Constraint<RequiresOpenshift> {
 
     @Override
     public void check(RequiresOpenshift context) throws UnsatisfiedRequirementException {
-        KubernetesClient client = new DefaultKubernetesClient(getConfig());
+        final List<String> extension = Arrays.asList(KUBERNETES_EXTENSION_NAME, OPENSHIFT_EXTENSION_NAME);
+
+        final DefaultConfiguration config = new ExtensionRegistrar().loadExtension(extension);
+
+        KubernetesClient client = new DefaultKubernetesClient(new ClientConfigBuilder().configuration(config).build());
 
         OkHttpClient httpClient = client.adapt(OkHttpClient.class);
         Request versionRequest = new Request.Builder()
@@ -44,18 +46,5 @@ public class OpenshiftRequirement implements Constraint<RequiresOpenshift> {
         } catch (IOException e) {
             throw new UnsatisfiedRequirementException("Error while checking Openshift version: [" + e.getMessage() + "]");
         }
-    }
-
-    private Config getConfig() {
-        final ConfigurationRegistrar configurationRegistrar = new ConfigurationRegistrar();
-        final ArquillianDescriptor arquillian = configurationRegistrar.loadConfiguration();
-
-        Map<String, String> map = new HashMap<>();
-        map.putAll(arquillian.extension(KUBERNETES_EXTENSION_NAME).getExtensionProperties());
-        map.putAll(arquillian.extension(OPENSHIFT_EXTENSION_NAME).getExtensionProperties());
-
-        final CubeOpenShiftConfiguration config = CubeOpenShiftConfiguration.fromMap(map);
-
-        return getConfigBuilder(config).build();
     }
 }
